Refactor integration tests to use objectProvider objects
